rename features/ops
see issue #80
