+++ src/main/java/net/imagej/ops/features/geometric/DefaultCircularityFeature.java
@@ -2,8 +2,8 @@ package net.imagej.ops.features.geometric;
 
 import net.imagej.ops.Op;
 import net.imagej.ops.features.FeatureService;
-import net.imagej.ops.features.geometric.GeometricFeatures.ConvexityFeature;
-import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonConvexHullPerimeterProvider;
+import net.imagej.ops.features.geometric.GeometricFeatures.CircularityFeature;
+import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonAreaProvider;
 import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonPerimeterProvider;
 
 import org.scijava.ItemIO;
@@ -11,19 +11,19 @@ import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link ConvexityFeature}. Use
+ * Generic implementation of {@link CircularityFeature}. Use
  * {@link FeatureService} to compile this {@link Op}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = Op.class, name = ConvexityFeature.NAME)
-public class DefConvexity implements ConvexityFeature {
+@Plugin(type = Op.class, name = CircularityFeature.NAME)
+public class DefaultCircularityFeature implements CircularityFeature {
+
+	@Parameter(type = ItemIO.INPUT)
+	private PolygonAreaProvider area;
 
 	@Parameter(type = ItemIO.INPUT)
 	private PolygonPerimeterProvider perimter;
-
-	@Parameter(type = ItemIO.INPUT)
-	private PolygonConvexHullPerimeterProvider convexHullPerimeter;
 
 	@Parameter(type = ItemIO.OUTPUT)
 	private double out;
@@ -35,8 +35,10 @@ public class DefConvexity implements ConvexityFeature {
 
 	@Override
 	public void run() {
-		out = convexHullPerimeter.getFeatureValue()
-				/ perimter.getFeatureValue();
+		out = 4
+				* Math.PI
+				* (area.getFeatureValue() / Math.pow(
+						perimter.getFeatureValue(), 2));
 	}
 
 }
