+++ src/main/java/net/imagej/ops/features/geometric/DefaultMajorAxisFeature.java
@@ -2,28 +2,24 @@ package net.imagej.ops.features.geometric;
 
 import net.imagej.ops.Op;
 import net.imagej.ops.features.FeatureService;
-import net.imagej.ops.features.geometric.GeometricFeatures.EccentricityFeature;
 import net.imagej.ops.features.geometric.GeometricFeatures.MajorAxisFeature;
-import net.imagej.ops.features.geometric.GeometricFeatures.MinorAxisFeature;
+import net.imagej.ops.features.geometric.helper.polygonhelper.MinorMajorAxisProvider;
 
 import org.scijava.ItemIO;
 import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link EccentricityFeature}. Use
+ * Generic implementation of {@link MajorAxisFeature}. Use
  * {@link FeatureService} to compile this {@link Op}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = Op.class, name = EccentricityFeature.NAME)
-public class DefEccentricity implements EccentricityFeature {
+@Plugin(type = Op.class, name = MajorAxisFeature.NAME)
+public class DefaultMajorAxisFeature implements MajorAxisFeature {
 
 	@Parameter(type = ItemIO.INPUT)
-	private MajorAxisFeature majorAxis;
-
-	@Parameter(type = ItemIO.INPUT)
-	private MinorAxisFeature minorAxis;
+	private MinorMajorAxisProvider axisProvider;
 
 	@Parameter(type = ItemIO.OUTPUT)
 	private double out;
@@ -35,6 +31,7 @@ public class DefEccentricity implements EccentricityFeature {
 
 	@Override
 	public void run() {
-		out = majorAxis.getFeatureValue() / minorAxis.getFeatureValue();
+		out = axisProvider.getOutput().getB();
 	}
+
 }
