+++ src/main/java/net/imagej/ops/features/geometric/DefaultMinorAxisFeature.java
@@ -2,7 +2,7 @@ package net.imagej.ops.features.geometric;
 
 import net.imagej.ops.Op;
 import net.imagej.ops.features.FeatureService;
-import net.imagej.ops.features.geometric.GeometricFeatures.MajorAxisFeature;
+import net.imagej.ops.features.geometric.GeometricFeatures.MinorAxisFeature;
 import net.imagej.ops.features.geometric.helper.polygonhelper.MinorMajorAxisProvider;
 
 import org.scijava.ItemIO;
@@ -10,13 +10,13 @@ import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link MajorAxisFeature}. Use
- * {@link FeatureService} to compile this {@link Op}.
+ * Generic implementation of {@link MinorAxisFeature}. Use {@link FeatureService} to
+ * compile this {@link Op}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = Op.class, name = MajorAxisFeature.NAME)
-public class DefMajorAxis implements MajorAxisFeature {
+@Plugin(type = Op.class, name = MinorAxisFeature.NAME)
+public class DefaultMinorAxisFeature implements MinorAxisFeature {
 
 	@Parameter(type = ItemIO.INPUT)
 	private MinorMajorAxisProvider axisProvider;
@@ -31,7 +31,7 @@ public class DefMajorAxis implements MajorAxisFeature {
 
 	@Override
 	public void run() {
-		out = axisProvider.getOutput().getB();
+		out = axisProvider.getOutput().getA();
 	}
 
 }
