+++ src/main/java/net/imagej/ops/features/geometric/DefaultRugosityFeature.java
@@ -2,8 +2,9 @@ package net.imagej.ops.features.geometric;
 
 import net.imagej.ops.Op;
 import net.imagej.ops.features.FeatureService;
-import net.imagej.ops.features.geometric.GeometricFeatures.CircularityFeature;
-import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonAreaProvider;
+import net.imagej.ops.features.geometric.GeometricFeatures.ConvexityFeature;
+import net.imagej.ops.features.geometric.GeometricFeatures.RugosityFeature;
+import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonConvexHullPerimeterProvider;
 import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonPerimeterProvider;
 
 import org.scijava.ItemIO;
@@ -11,19 +12,19 @@ import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link CircularityFeature}. Use
+ * Generic implementation of {@link ConvexityFeature}. Use
  * {@link FeatureService} to compile this {@link Op}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = Op.class, name = CircularityFeature.NAME)
-public class DefCircularity implements CircularityFeature {
-
-	@Parameter(type = ItemIO.INPUT)
-	private PolygonAreaProvider area;
+@Plugin(type = Op.class, name = RugosityFeature.NAME)
+public class DefaultRugosityFeature implements RugosityFeature {
 
 	@Parameter(type = ItemIO.INPUT)
 	private PolygonPerimeterProvider perimter;
+
+	@Parameter(type = ItemIO.INPUT)
+	private PolygonConvexHullPerimeterProvider convexHullPerimeter;
 
 	@Parameter(type = ItemIO.OUTPUT)
 	private double out;
@@ -35,10 +36,8 @@ public class DefCircularity implements CircularityFeature {
 
 	@Override
 	public void run() {
-		out = 4
-				* Math.PI
-				* (area.getFeatureValue() / Math.pow(
-						perimter.getFeatureValue(), 2));
+		out = perimter.getFeatureValue()
+				/ convexHullPerimeter.getFeatureValue();
 	}
 
 }
