+++ src/main/java/net/imagej/ops/features/geometric/DefaultSolidityFeature.java
@@ -2,28 +2,28 @@ package net.imagej.ops.features.geometric;
 
 import net.imagej.ops.Op;
 import net.imagej.ops.features.FeatureService;
-import net.imagej.ops.features.geometric.GeometricFeatures.RectangularityFeature;
+import net.imagej.ops.features.geometric.GeometricFeatures.SolidityFeature;
 import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonAreaProvider;
-import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonSmallestEnclosingRectangleAreaProvider;
+import net.imagej.ops.features.geometric.helper.polygonhelper.PolygonConvexHullAreaProvider;
 
 import org.scijava.ItemIO;
 import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link RectangularityFeature}. Use
+ * Generic implementation of {@link SolidityFeature}. Use
  * {@link FeatureService} to compile this {@link Op}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = Op.class, name = RectangularityFeature.NAME)
-public class DefRectangularity implements RectangularityFeature {
+@Plugin(type = Op.class, name = SolidityFeature.NAME)
+public class DefaultSolidityFeature implements SolidityFeature {
 
 	@Parameter(type = ItemIO.INPUT)
 	private PolygonAreaProvider area;
 
 	@Parameter(type = ItemIO.INPUT)
-	private PolygonSmallestEnclosingRectangleAreaProvider serArea;
+	private PolygonConvexHullAreaProvider convexHullArea;
 
 	@Parameter(type = ItemIO.OUTPUT)
 	private double out;
@@ -35,7 +35,7 @@ public class DefRectangularity implements RectangularityFeature {
 
 	@Override
 	public void run() {
-		out = area.getFeatureValue() / serArea.getFeatureValue();
+		out = area.getFeatureValue() / convexHullArea.getFeatureValue();
 	}
 
 }
