+++ src/main/java/net/imagej/ops/features/sets/GeometricFeatureSet.java
@@ -32,7 +32,7 @@ package net.imagej.ops.features.sets;
 
 import net.imagej.ops.features.AbstractFeatureSet;
 import net.imagej.ops.features.FeatureSet;
-import net.imagej.ops.features.geometric.DefRoundness;
+import net.imagej.ops.features.geometric.DefaultRoundnessFeature;
 import net.imagej.ops.features.geometric.GeometricFeatures.AreaFeature;
 import net.imagej.ops.features.geometric.GeometricFeatures.CircularityFeature;
 import net.imagej.ops.features.geometric.GeometricFeatures.ConvexityFeature;
